+++ LICENSE.txt
@@ -1,4 +1,4 @@
-Copyright © 2014-2015 Cask Data, Inc.
+Copyright © 2014-2019 Cask Data, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
