+++ src/main/java/io/cdap/plugin/tokenizer/Tokenizer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2016 Cask Data, Inc.
+ * Copyright © 2016-2019 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -13,24 +13,25 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package co.cask.hydrator.plugin;
+package io.cdap.plugin.tokenizer;
 
-import co.cask.cdap.api.annotation.Description;
-import co.cask.cdap.api.annotation.Name;
-import co.cask.cdap.api.annotation.Plugin;
-import co.cask.cdap.api.data.format.StructuredRecord;
-import co.cask.cdap.api.data.schema.Schema;
-import co.cask.cdap.api.plugin.PluginConfig;
-import co.cask.cdap.etl.api.PipelineConfigurer;
-import co.cask.cdap.etl.api.batch.SparkCompute;
-import co.cask.cdap.etl.api.batch.SparkExecutionPluginContext;
-import co.cask.cdap.format.StructuredRecordStringConverter;
 import com.google.common.base.Preconditions;
+import io.cdap.cdap.api.annotation.Description;
+import io.cdap.cdap.api.annotation.Name;
+import io.cdap.cdap.api.annotation.Plugin;
+import io.cdap.cdap.api.data.format.StructuredRecord;
+import io.cdap.cdap.api.data.schema.Schema;
+import io.cdap.cdap.api.plugin.PluginConfig;
+import io.cdap.cdap.etl.api.PipelineConfigurer;
+import io.cdap.cdap.etl.api.batch.SparkCompute;
+import io.cdap.cdap.etl.api.batch.SparkExecutionPluginContext;
+import io.cdap.cdap.format.StructuredRecordStringConverter;
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.api.java.function.Function;
 import org.apache.spark.ml.feature.RegexTokenizer;
-import org.apache.spark.sql.DataFrame;
+import org.apache.spark.sql.Dataset;
+import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SQLContext;
 
 import java.util.ArrayList;
@@ -75,7 +76,7 @@ public class Tokenizer extends SparkCompute<StructuredRecord, StructuredRecord> 
     JavaSparkContext javaSparkContext = context.getSparkContext();
     SQLContext sqlContext = new SQLContext(javaSparkContext);
     if (input == null) {
-      return context.getSparkContext().emptyRDD();
+      return context.getSparkContext().<StructuredRecord>emptyRDD();
     }
     outputSchema = outputSchema != null ? outputSchema : config.getOutputSchema(input.first().getSchema(),
                                                                                 config.outputColumn);
@@ -85,11 +86,11 @@ public class Tokenizer extends SparkCompute<StructuredRecord, StructuredRecord> 
         return StructuredRecordStringConverter.toJsonString(structuredRecord);
       }
     });
-    DataFrame sentenceDataFrame = sqlContext.read().json(javardd);
+    Dataset<Row> sentenceDataFrame = sqlContext.read().json(javardd);
     RegexTokenizer regexTokenizer = new RegexTokenizer().setInputCol(config.columnToBeTokenized)
       .setOutputCol(config.outputColumn)
       .setPattern(config.patternSeparator);
-    DataFrame tokenizedDataFrame = regexTokenizer.transform(sentenceDataFrame);
+    Dataset<Row> tokenizedDataFrame = regexTokenizer.transform(sentenceDataFrame);
     JavaRDD<StructuredRecord> output = tokenizedDataFrame.toJSON().toJavaRDD()
       .map(new Function<String, StructuredRecord>() {
         @Override
